Sparse: fixing a few issues related to coo2csr and par_ilut benchmark
coo2csr needed to be split between src and impl, eventually we could
decide to add a spec file and do the whole ETI for it, to be discussed.
Additionally coo2csr uses features that are in Kokkos Core develop
which prevents Kokkos Kernels from building against 4.0.00 or 4.0.01.
Adding preprocessor checks to avoid this issue.

The par_ilut benchmark unfortunately does not make any sense, it
force its way around using google benchmark for every things from
arguments setting, to timing and to printing outputs. This leads
to unsued variable issues which cannot be fixed... This should really
be reverted but I am just commenting it out for now.
